+++ src/app/main/package-selection/package-details/package-details.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
-import { Package } from '../../../core/models/package.model';
+import { Package, PackageVersion } from '../../../core/models/package.model';
 import { PreferencesService } from '../../../core/services/preferences.service';
 import { Subscription } from 'rxjs';
 import { PackageService } from '../../../core/services/package.service';
@@ -11,6 +11,7 @@ import { PackageService } from '../../../core/services/package.service';
 })
 export class PackageDetailsComponent implements OnInit, OnDestroy {
   package: Package;
+  packageVersion: PackageVersion;
 
   shouldHumanize = this.prefs.get('humanizePackageNames');
 
@@ -31,6 +32,7 @@ export class PackageDetailsComponent implements OnInit, OnDestroy {
     this.subscription.add(
       this.packages.selectedPackage.subscribe(pkg => {
         this.package = pkg;
+        this.packageVersion = pkg && pkg.latestVersion;
       })
     );
   }